Get datasets from Vancouver Open Data Portal

Get datasets from Vancouver Open Data Portal

get_cov_data(
  dataset_id,
  format = c("csv", "geojson"),
  select = "*",
  where = NULL,
  apikey = getOption("VancouverOpenDataApiKey"),
  rows = NULL,
  cast_types = TRUE,
  refresh = FALSE
)

Arguments

dataset_id
Dataset id from the Vancouver Open Data catalogue

format
`csv` or `geojson` are supported at this time (default `csv`)

select
select string for fields to return, returns all fields by default. It accepts ODSQL syntax.

where
Query parameter to filter data (default `NULL` no filter) It accepts ODSQL syntax.

apikey
Vancouver Open Data API key, default `getOption("VancouverOpenDataApiKey")`

rows
Maximum number of rows to return (default `NULL` returns all rows)

cast_types
Logical, use metadata to look up types and type-cast automatically, default `TRUE`

refresh
refresh cached data, default `FALSE``

Value

tibble or sf object data table output, depending on the value of the `format` parameter

Examples

# Get all parking tickets issued at the 1100 block of Alberni Street between 2017 and 2019
if (FALSE) {
get_cov_data("parking-tickets-2017-2019",where = "block = 1100 AND street = 'ALBERNI ST'")
}