Get metadata for CoV open data dataset

Get metadata for CoV open data dataset

get_cov_metadata(
  dataset_id,
  apikey = getOption("VancouverOpenDataApiKey"),
  refresh = FALSE
)

Arguments

dataset_id
the CoV open data dataset id

apikey
the CoV open data API key, optional

refresh
refresh cached data, default `FALSE``

Value

tibble format data table output

Examples

# Get the metadata for the street trees dataset
get_cov_metadata("street-trees")

#> # A tibble: 19 x 4
#>    name         type    label       description                                 
#>    <chr>        <chr>   <chr>       <chr>                                       
#>  1 tree_id      text    TREE_ID     Numerical ID                                
#>  2 civic_number text    CIVIC_NUMB… Street address of the site at which the tre…
#>  3 std_street   text    STD_STREET  Street name of the site at which the tree i…
#>  4 genus_name   text    GENUS_NAME  Genus name                                  
#>  5 species_name text    SPECIES_NA… Species name                                
#>  6 cultivar_na… text    CULTIVAR_N… Cultivar name                               
#>  7 common_name  text    COMMON_NAME Common name                                 
#>  8 assigned     text    ASSIGNED    Indicates whether the address is made up to…
#>  9 root_barrier text    ROOT_BARRI… Root barrier installed (Y = Yes, N = No)    
#> 10 plant_area   text    PLANT_AREA  B = behind sidewalk, G = in tree grate, N =…
#> 11 on_street_b… text    ON_STREET_… The street block at which the tree is physi…
#> 12 on_street    text    ON_STREET   The name of the street at which the tree is…
#> 13 neighbourho… text    NEIGHBOURH… City's defined local area in which the tree…
#> 14 street_side… text    STREET_SID… The street side which the tree is physicall…
#> 15 height_rang… int     HEIGHT_RAN… 0-10 for every 10 feet (e.g., 0 = 0-10 ft, …
#> 16 diameter     double  DIAMETER    DBH in inches (DBH stands for diameter of t…
#> 17 curb         text    CURB        Curb presence (Y = Yes, N = No)             
#> 18 date_planted date    DATE_PLANT… The date of planting in YYYYMMDD format.  D…
#> 19 geom         geo_sh… Geom        Spatial representation of feature